moves response to a better directory
